requirements.txt
# Requirements - I had to run "sudo apt-get install libpq-dev" before running pip install
# Also install libjpeg-dev before pillow sudo apt-get install libjpeg-dev
# reinstall pillow: pip install --no-cache-dir -I pillow
gunicorn==20.0.4
Django==1.11.29
Pillow==2.5.1
psycopg2==2.8.3
#bootstrap-admin  # https://github.com/douglasmiranda/django-admin-bootstrap
ConcurrentLogHandler==0.9.1
reportlab==3.3.0
django-filter==1.1.0
djangorestframework==3.4.6
django-cors-headers==1.3.1

# AFIP
httplib2==0.9.2
M2Crypto==0.36.0
git+https://github.com/pysimplesoap/pysimplesoap.git@stable_py3k
git+https://github.com/reingart/pyafipws.git@py3k

# testing or dev tools (TODO: should go on a different file)
mock==3.0.5                    # Mocking and patching library for testing
pylint
pylint-django==0.11.1
pytest
pytest-django
xlrd # Tool informe contadora